airport_traffic_app/requirements.txt
flask
requests
pandas
pytest
python-dotenv
pyopensky
